Updated scripts and directories.
Live cd's are built by using live-build to build an up-to-date vanilla kernel.
Then, extract the image, chroot into, and install the rt kernel + rtxi. The
process is time-consuming, but this has by far produced the most consistent
results. The config options for live-build and the list of packages to be
installed on the cd are in their own directories.